% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKET.R
\name{kernTEST}
\alias{kernTEST}
\title{kernTEST}
\usage{
kernTEST(
  RESI,
  KK,
  OMNI,
  nPERMS = 1e+05,
  iter1 = 50,
  iter2 = 1000,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{RESI}{A numeric vector of null model residuals
\code{names(RESI)} must be set to maintain sample ordering}

\item{KK}{An array containing double-centered positive semi-definite
kernel matrices. Refer to \code{MiRKAT::D2K()} for transforming 
distance matrices to kernel matrices. The \code{dimnames(KK)[[1]]} and 
\code{dimnames(KK)[[2]]} must match \code{names(RESI)}.
Also set dimnames(KK)[[3]] to keep track of each kernel matrix.}

\item{OMNI}{A matrix of zeros and ones. Each column corresponds to a
distance matrix while each row corresponds to an omnibus test. Set
\code{rownames(OMNI)} for labeling outputted p-values and 
\code{colnames(OMNI)} which should match \code{dimnames(KK)[[3]]}.}

\item{nPERMS}{A positive integer to specify the number of
permutation-based p-value calculation}

\item{iter1}{A positive integer for displaying a dot for every
'iter1' iterations through the loop's progress}

\item{iter2}{A positive integer for displaying a newline for every
'iter2' iterations through the loop's progress}

\item{ncores}{A positive integer for the number of cores/threads
to reduce computational runtime when running for loops}

\item{verbose}{Boolean value to display verbose function output.}
}
\description{
kernTEST
}
