% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{kOT_sim_AGG}
\alias{kOT_sim_AGG}
\title{kOT_sim_AGG}
\usage{
kOT_sim_AGG(work_dir)
}
\arguments{
\item{work_dir}{A full path to create "sim_ROKET" and subdirectories}
}
\value{
Nothing. Png files are created within the simulation ROKET directory.
}
\description{
kOT_sim_AGG
}
