% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKET.R
\name{kOT_sim_make}
\alias{kOT_sim_make}
\title{kOT_sim_make}
\usage{
kOT_sim_make(work_dir, NN = 200, nGENE = 500, nPATH = 25, RR = 200)
}
\arguments{
\item{work_dir}{A full path to create "sim_ROKET" and subdirectories}

\item{NN}{A positive integer for sample size}

\item{nGENE}{A positive integer for number of genes to simulate}

\item{nPATH}{A positive integer for number of pathways to simulate}

\item{RR}{A positive integer for number of replicates to simulate}
}
\description{
Generates simulation files
}
