% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKET.R
\name{kOT_sim_OT}
\alias{kOT_sim_OT}
\title{kOT_sim_OT}
\usage{
kOT_sim_OT(work_dir, NN, nGENE, nPATH, SCEN, ncores = 1)
}
\arguments{
\item{work_dir}{A full path to create "sim_ROKET" and subdirectories}

\item{NN}{A positive integer for sample size}

\item{nGENE}{A positive integer for number of genes to simulate}

\item{nPATH}{A positive integer for number of pathways to simulate}

\item{SCEN}{An integer taking values 1, 2, 3, or 4}

\item{ncores}{A positive integer specifying the number of
cores/threads to use for optimal transport calculations}
}
\description{
kOT_sim_OT
}
