% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{kOT_sim_REG}
\alias{kOT_sim_REG}
\title{kOT_sim_REG}
\usage{
kOT_sim_REG(work_dir, NN, nGENE, nPATH, SCEN, rr)
}
\arguments{
\item{work_dir}{A full path to create "sim_ROKET" and subdirectories}

\item{NN}{A positive integer for sample size}

\item{nGENE}{A positive integer for number of genes to simulate}

\item{nPATH}{A positive integer for number of pathways to simulate}

\item{SCEN}{An integer taking values 1, 2, 3, or 4}

\item{rr}{A positive integer indexing a replicate}
}
\value{
Nothing. A rds file is created within the simulation ROKET directory.
}
\description{
kOT_sim_REG
}
