% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROKET.R
\name{kernTEST}
\alias{kernTEST}
\title{kernTEST}
\usage{
kernTEST(
  RESI,
  KK,
  nPERMS = 1e+05,
  iter1 = 50,
  iter2 = 1000,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{RESI}{A numeric vector of null model residuals
\code{names(RESI)} must be set to maintain sample ordering}

\item{KK}{A list containing double-centered positive semi-definite
kernel matrices. Refer to \code{MiRKAT::D2K()} for transforming 
distance matrices to kernel matrices. The \code{colnames()} and 
\code{rownames()} per kernel matrix must match \code{names(RESI)}.
Also set names(KK) to keep track of each kernel matrix.}

\item{nPERMS}{A positive integer to specify the number of
permutation-based p-value calculation}

\item{iter1}{A positive integer for displaying a dot for every
'iter1' iterations through the loop's progress}

\item{iter2}{A positive integer for displaying a newline for every
'iter2' iterations through the loop's progress}

\item{ncores}{A positive integer for the number of cores/threads
to reduce computational runtime when running for loops}

\item{verbose}{Boolean set to TRUE to display verbose progress output}
}
\description{
kernTEST
}
